arch_installation, arch_configuration: Added new notes for these
Added notes for arch installation and arch configuration (some of them
still need to be tested). Also, org-babel may be used to tangle the
configuration one to create a script that configures part of the
system (this is really useful after a fresh install). More things
still need to be added but this is a great start.